+++ src/core/storage/index.ts
@@ -1,1 +1,2 @@
 export * from "./StorageAdapter";
+export * from "./StorageAdapterProvider";